Create SerializedSuiteController (#987)
