+++ src/Controller/SuiteController.php
@@ -4,28 +4,20 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
-use App\Entity\SerializedSuite;
 use App\Entity\Suite;
 use App\Exception\EmptyEntityIdException;
 use App\Exception\InvalidRequestException;
 use App\Exception\ModifyReadOnlyEntityException;
 use App\Exception\NonUniqueEntityLabelException;
-use App\Message\SerializeSuite;
 use App\Repository\SuiteRepository;
 use App\Request\SuiteRequest;
-use App\Response\YamlResponse;
 use App\Security\EntityAccessChecker;
 use App\Services\ExceptionFactory;
 use App\Services\Suite\Factory;
 use App\Services\Suite\Mutator;
-use App\Services\Suite\SerializedSuiteFactory;
-use App\Services\SuiteSerializer;
-use League\Flysystem\FilesystemException;
 use SmartAssert\UsersSecurityBundle\Security\User;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
@@ -45,7 +37,7 @@ class SuiteController
      * @throws EmptyEntityIdException
      * @throws InvalidRequestException
      */
-    #[Route(SuiteRoutes::ROUTE_SUITE_BASE, name: 'user_suite_create', methods: ['POST'])]
+    #[Route(SuiteRoutes::ROUTE_SUITE_BASE, name: 'suite_create', methods: ['POST'])]
     public function create(SuiteRequest $request): Response
     {
         $this->entityAccessChecker->denyAccessUnlessGranted($request->source);
@@ -64,7 +56,7 @@ class SuiteController
     /**
      * @throws AccessDeniedException
      */
-    #[Route(SuiteRoutes::ROUTE_SUITE, name: 'user_suite_get', methods: ['GET'])]
+    #[Route(SuiteRoutes::ROUTE_SUITE, name: 'suite_get', methods: ['GET'])]
     public function get(Suite $suite): Response
     {
         $this->entityAccessChecker->denyAccessUnlessGranted($suite);
@@ -75,7 +67,7 @@ class SuiteController
     /**
      * @throws AccessDeniedException
      */
-    #[Route(SuiteRoutes::ROUTE_SUITES, name: 'user_suite_list', methods: ['GET'])]
+    #[Route(SuiteRoutes::ROUTE_SUITES, name: 'suite_list', methods: ['GET'])]
     public function list(User $user): Response
     {
         $suites = $this->repository->findForUser($user);
@@ -88,7 +80,7 @@ class SuiteController
      * @throws InvalidRequestException
      * @throws ModifyReadOnlyEntityException
      */
-    #[Route(SuiteRoutes::ROUTE_SUITE, name: 'user_suite_update', methods: ['POST'])]
+    #[Route(SuiteRoutes::ROUTE_SUITE, name: 'suite_update', methods: ['POST'])]
     public function update(Suite $suite, SuiteRequest $request): Response
     {
         $this->entityAccessChecker->denyAccessUnlessGranted($suite);
@@ -111,7 +103,7 @@ class SuiteController
     /**
      * @throws AccessDeniedException
      */
-    #[Route(SuiteRoutes::ROUTE_SUITE, name: 'user_suite_delete', methods: ['DELETE'])]
+    #[Route(SuiteRoutes::ROUTE_SUITE, name: 'suite_delete', methods: ['DELETE'])]
     public function delete(Suite $suite): Response
     {
         $this->entityAccessChecker->denyAccessUnlessGranted($suite);
@@ -121,37 +113,5 @@ class SuiteController
         }
 
         return new JsonResponse($suite);
-    }
-
-    /**
-     * @throws AccessDeniedException
-     * @throws FilesystemException
-     */
-    #[Route(SuiteRoutes::ROUTE_SUITE . '/read', name: 'user_suite_read', methods: ['GET'])]
-    public function read(SerializedSuite $serializedSuite, SuiteSerializer $suiteSerializer): Response
-    {
-        $this->entityAccessChecker->denyAccessUnlessGranted($serializedSuite);
-
-        return new YamlResponse($suiteSerializer->read($serializedSuite));
-    }
-
-    /**
-     * @throws AccessDeniedException
-     * @throws EmptyEntityIdException
-     */
-    #[Route(SuiteRoutes::ROUTE_SUITE . '/serialize', name: 'user_suite_serialize', methods: ['POST'])]
-    public function serialize(
-        Request $request,
-        Suite $suite,
-        MessageBusInterface $messageBus,
-        SerializedSuiteFactory $serializedSuiteFactory,
-    ): Response {
-        $this->entityAccessChecker->denyAccessUnlessGranted($suite);
-
-        $serializedSuite = $serializedSuiteFactory->create($suite, $request);
-
-        $messageBus->dispatch(SerializeSuite::createFromSerializedSuite($serializedSuite));
-
-        return new JsonResponse($serializedSuite, 202);
     }
 }
